Update index.css
Made things look pretty. Content is coming soon. I'd like for all of the websites content to be accessible from one page (Good ol' AJAX). Animations and images will be added soon as well. 

Also considering experimenting with a terminal "look and feel" for the welcome page with a sort of hacker-esque vibe.